SawTooth and Triangle Wave Signals with frequency of 83 Hz
Need to reduce the frequency down to 65 Hz. This will be possible by adding near to microseconds of delay in each for loop.

Period --> 12 ms
Required Period ~ 15 ms
∆Period --> 3 ms
Sample Delay --> 3 ms / 0xffff = 45.77 ns of delay

Adding ns of delay is impossible in the board so either add 3 ms of delay after each period or add 3 ms of delay at the peak amplitude of each waves